+++ pos_customer_required/__init__.py
