+++ pos_customer_required/demo/pos_config.yml
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2016-Today: La Louve (<http://www.lalouve.net/>)
+# @author: Sylvain LE GAL (https://twitter.com/legalsylvain)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+- !record {model: pos.config, id: point_of_sale.pos_config_main}:
+    require_customer: order
