+++ pos_customer_required/i18n/fr.po
@@ -1,0 +1,71 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+#	* pos_customer_required
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 9.0c\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2016-04-03 23:56+0000\n"
+"PO-Revision-Date: 2016-04-03 23:56+0000\n"
+"Last-Translator: <>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: pos_customer_required
+#: selection:pos.config,require_customer:0
+msgid "Optional"
+msgstr "Facultatif"
+
+#. module: pos_customer_required
+#: model:ir.model,name:pos_customer_required.model_pos_order
+msgid "Point of Sale"
+msgstr "Point de vente"
+
+#. module: pos_customer_required
+#: model:ir.model,name:pos_customer_required.model_pos_make_payment
+msgid "Point of Sale Payment"
+msgstr "Paiement du ticket"
+
+#. module: pos_customer_required
+#: model:ir.model.fields,field_description:pos_customer_required.field_pos_config_require_customer
+msgid "Require Customer"
+msgstr "Client requis"
+
+#. module: pos_customer_required
+#: model:ir.model.fields,field_description:pos_customer_required.field_pos_order_require_customer
+msgid "Require customer"
+msgstr "Client requis"
+
+#. module: pos_customer_required
+#: model:ir.model.fields,help:pos_customer_required.field_pos_config_require_customer
+msgid "Require customer for orders in this point of sale:\n"
+"* 'Optional' (customer is optional);* 'Required before paying';* 'Required before starting the order';"
+msgstr "Client requis pour vendre dans le point de vente:\n"
+"* 'Optional' (le client est facultatif);\n* 'Requis avant de payer';\n* 'Requis avant de commencer la vente';"
+
+#. module: pos_customer_required
+#: selection:pos.config,require_customer:0
+msgid "Required before paying"
+msgstr "Requis avant de payer"
+
+#. module: pos_customer_required
+#: selection:pos.config,require_customer:0
+msgid "Required before starting the order"
+msgstr "Requis avant de commencer la vente"
+
+#. module: pos_customer_required
+#: model:ir.model.fields,help:pos_customer_required.field_pos_order_require_customer
+msgid "True if a customer is required to begin the order.\n"
+"See the PoS Config to change this setting"
+msgstr "Vrai si un client est requis pour commencer une vente.\n"
+"Voir la configuration du point de vente pour changer ce paramètre."
+
+#. module: pos_customer_required
+#: model:ir.model,name:pos_customer_required.model_pos_config
+msgid "pos.config"
+msgstr "pos.config"
+
