+++ pos_customer_required/models/__init__.py
