+++ pos_customer_required/models/pos_make_payment.py
@@ -1,13 +1,8 @@
 # -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Copyright (C) 2016-Today La Louve (<http://www.lalouve.net/>)
-#
-#    @author: Sylvain LE GAL (https://twitter.com/legalsylvain)
-#
-#    The licence is in the file __openerp__.py
-#
-##############################################################################
+# Copyright (C) 2016-Today: La Louve (<http://www.lalouve.net/>)
+# @author: Sylvain LE GAL (https://twitter.com/legalsylvain)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
 
 from openerp import models, api, _
 from openerp.exceptions import UserError
@@ -23,7 +18,8 @@ class PosMakePayment(models.TransientModel):
         order = order_obj.browse(self.env.context.get('active_id', False))
 
         # Check if control is required
-        if order and order.session_id.config_id.require_customer != 'no':
+        if not order.partner_id\
+                and order.session_id.config_id.require_customer != 'no':
             raise UserError(_(
                 "An anonymous order cannot be confirmed.\n"
                 "Please select a customer for this order."))
