+++ pos_customer_required/models/pos_order.py
@@ -1,15 +1,9 @@
 # -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Copyright (C) 2004-Today Apertoso NV (<http://www.apertoso.be>)
-#    Copyright (C) 2016-Today La Louve (<http://www.lalouve.net/>)
-#
-#    @author: Jos DE GRAEVE (<Jos.DeGraeve@apertoso.be>)
-#    @author: Sylvain LE GAL (https://twitter.com/legalsylvain)
-#
-#    The licence is in the file __openerp__.py
-#
-##############################################################################
+# Copyright (C) 2004-Today Apertoso NV (<http://www.apertoso.be>)
+# Copyright (C) 2016-Today: La Louve (<http://www.lalouve.net/>)
+# @author: Jos DE GRAEVE (<Jos.DeGraeve@apertoso.be>)
+# @author: Sylvain LE GAL (https://twitter.com/legalsylvain)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 from openerp import fields, models, exceptions, api
 from openerp.tools.translate import _
@@ -18,11 +12,12 @@ from openerp.tools.translate import _
 class PosOrder(models.Model):
     _inherit = 'pos.order'
 
+    @api.multi
     @api.depends('session_id.config_id.require_customer')
     def compute_require_customer(self):
         for order in self:
-            self.require_customer = (
-                self.session_id.config_id.require_customer == 'order')
+            order.require_customer = (
+                order.session_id.config_id.require_customer == 'order')
 
     require_customer = fields.Boolean(
         compute='compute_require_customer', string='Require customer',
