+++ pos_customer_required/static/src/js/pos_customer_required.js
@@ -9,7 +9,7 @@
 */
 
 
-odoo.define('pos_require_customer.pos_require_customer', function (require) {
+odoo.define('pos_customer_required.pos_customer_required', function (require) {
     "use strict";
 
     var screens = require('point_of_sale.screens');
