+++ pos_customer_required/static/src/xml/templates.xml
@@ -13,7 +13,7 @@
     <data>
         <template id="assets_frontend" inherit_id="point_of_sale.assets">
             <xpath expr="." position="inside">
-                <script type="text/javascript" src="/pos_require_customer/static/src/js/pos_require_customer.js"/>
+                <script type="text/javascript" src="/pos_customer_required/static/src/js/pos_customer_required.js"/>
             </xpath>
         </template>
     </data>
