+++ pos_customer_required/tests/__init__.py
