+++ pos_customer_required/views/pos_config_view.xml
