+++ pos_customer_required/views/pos_order_view.xml
