--- pos_require_customer/demo/pos_config.yml
@@ -1,13 +1,0 @@
-# -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Copyright (C) 2016-Today La Louve (<http://www.lalouve.net/>)
-#
-#    @author: Sylvain LE GAL (https://twitter.com/legalsylvain)
-#
-#    The licence is in the file __openerp__.py
-#
-##############################################################################
-
-- !record {model: pos.config, id: point_of_sale.pos_config_main}:
-    require_customer: order
